refactor: 로비와 대기방 DM 상태 관리를 공통 훅으로 분리
