+++ src/pages/lobby/LobbyPage.tsx
@@ -10,17 +10,8 @@ import {
 } from '../../components/header/profileMenu'
 import { UserListPanel } from '../../features/presence/components/UserListPanel'
 import { DirectMessagePanel } from '../../features/presence/components/DirectMessagePanel'
-import {
-  sendDirectMessage,
-  subscribeDirectMessageSocketEvents,
-} from '../../features/presence/directMessageSocket'
 import { useOnlineUsersSocket } from '../../features/presence/hooks'
-import { isDirectMessageAllowed } from '../../features/presence/status'
-import type {
-  DirectMessage,
-  DirectMessageReceiveSocketPayload,
-  OnlineUser,
-} from '../../features/presence/types'
+import { useDirectMessageController } from '../../features/presence/useDirectMessageController'
 import type { LobbyRoomFilter } from './api'
 import { useLobbyRoomsQuery } from './hooks'
 import { LobbyControls } from './LobbyControls'
@@ -48,14 +39,6 @@ function LobbyPage() {
   const [isUserListOpen, setIsUserListOpen] = useState(true)
   const [isCreateRoomModalOpen, setIsCreateRoomModalOpen] = useState(false)
   const [isCreateRoomPending, setIsCreateRoomPending] = useState(false)
-  const [dmTargetUser, setDmTargetUser] = useState<OnlineUser | null>(null)
-  const [directMessagesByUserId, setDirectMessagesByUserId] = useState<
-    Record<string, DirectMessage[]>
-  >({})
-  const [
-    unreadDirectMessageCountByUserId,
-    setUnreadDirectMessageCountByUserId,
-  ] = useState<Record<string, number>>({})
   const [selectedPrivateRoom, setSelectedPrivateRoom] =
     useState<LobbyRoom | null>(null)
   const [privateRoomPassword, setPrivateRoomPassword] = useState('')
@@ -94,86 +77,20 @@ function LobbyPage() {
     isError: isUsersError,
   } = useOnlineUsersSocket()
 
-  const openedDirectMessageUserId = dmTargetUser?.id
-
-  // DM 수신 이벤트를 구독해 대화 목록과 unread 카운트를 갱신
-  useEffect(() => {
-    // 비로그인 상태에서는 구독을 등록하지 않음
-    if (!session) {
-      return
-    }
-
-    const unsubscribe = subscribeDirectMessageSocketEvents({
-      onReceive: (payload: DirectMessageReceiveSocketPayload) => {
-        const receivedMessage: DirectMessage = {
-          id: `${payload.sender_id}-${payload.sent_at}`,
-          senderId: payload.sender_id,
-          senderNickname: payload.sender_nickname,
-          content: payload.message,
-          sentAt: payload.sent_at,
-        }
-
-        setDirectMessagesByUserId((previousMessagesByUserId) => {
-          const previousMessages =
-            previousMessagesByUserId[payload.sender_id] ?? []
-          const hasSameMessage = previousMessages.some(
-            (message) => message.id === receivedMessage.id
-          )
-
-          // 동일 메시지 ID는 중복 삽입을 방지
-          if (hasSameMessage) {
-            return previousMessagesByUserId
-          }
-
-          return {
-            ...previousMessagesByUserId,
-            [payload.sender_id]: [...previousMessages, receivedMessage],
-          }
-        })
-
-        // 현재 열려 있는 사용자 메시지는 unread 카운트에서 제외
-        if (payload.sender_id === openedDirectMessageUserId) {
-          return
-        }
-
-        setUnreadDirectMessageCountByUserId((previousCountByUserId) => {
-          const previousCount = previousCountByUserId[payload.sender_id] ?? 0
-
-          return {
-            ...previousCountByUserId,
-            [payload.sender_id]: previousCount + 1,
-          }
-        })
-      },
-    })
-
-    return () => {
-      unsubscribe()
-    }
-  }, [openedDirectMessageUserId, session])
-
-  // 접속자 목록 갱신 시 DM 대상 사용자 참조를 최신 객체로 동기화
-  useEffect(() => {
-    if (!dmTargetUser) {
-      return
-    }
-
-    const matchedUser = users.find((user) => user.id === dmTargetUser.id)
-
-    if (!matchedUser) {
-      setDmTargetUser(null)
-      return
-    }
-
-    if (!isDirectMessageAllowed(matchedUser.status)) {
-      setDmTargetUser(null)
-      return
-    }
-
-    if (matchedUser !== dmTargetUser) {
-      setDmTargetUser(matchedUser)
-    }
-  }, [dmTargetUser, users])
+  const {
+    dmTargetUser,
+    directMessagesByUserId,
+    unreadDirectMessageCountByUserId,
+    openDirectMessage,
+    closeDirectMessage,
+    sendDirectMessage,
+  } = useDirectMessageController({
+    session,
+    users,
+    onBlockedByPlaying: () => {
+      toast.error('게임중인 유저에게는 DM을 보낼 수 없습니다.')
+    },
+  })
 
   function handleLogout() {
     clearSession()
@@ -183,66 +100,6 @@ function LobbyPage() {
   // 프로필 드롭다운에서 마이페이지 이동 처리
   function handleGoMyPage() {
     navigate('/my-page')
-  }
-
-  // DM 창을 열고 해당 사용자 unread 카운트를 초기화
-  function handleOpenDirectMessage(user: OnlineUser) {
-    if (!isDirectMessageAllowed(user.status)) {
-      toast.error('게임중인 유저에게는 DM을 보낼 수 없습니다.')
-      return
-    }
-
-    setDmTargetUser(user)
-    setUnreadDirectMessageCountByUserId((previousCountByUserId) => {
-      if (!previousCountByUserId[user.id]) {
-        return previousCountByUserId
-      }
-
-      const nextCountByUserId = { ...previousCountByUserId }
-      delete nextCountByUserId[user.id]
-      return nextCountByUserId
-    })
-  }
-
-  function handleCloseDirectMessage() {
-    setDmTargetUser(null)
-  }
-
-  // 로컬 DM 목록에 메시지를 추가하고 소켓 전송 실행
-  function handleSendDirectMessage(message: string) {
-    // 대상 사용자 또는 세션이 없으면 전송 중단
-    if (!dmTargetUser || !session) {
-      return
-    }
-
-    if (!isDirectMessageAllowed(dmTargetUser.status)) {
-      toast.error('게임중인 유저에게는 DM을 보낼 수 없습니다.')
-      return
-    }
-
-    const targetUserId = dmTargetUser.id
-    const nextDirectMessage: DirectMessage = {
-      id: `${targetUserId}-${Date.now()}`,
-      senderId: session.userId,
-      senderNickname: session.nickname,
-      content: message,
-      sentAt: new Date().toISOString(),
-    }
-
-    setDirectMessagesByUserId((previousMessagesByUserId) => {
-      const previousMessages = previousMessagesByUserId[targetUserId] ?? []
-
-      return {
-        ...previousMessagesByUserId,
-        [targetUserId]: [...previousMessages, nextDirectMessage],
-      }
-    })
-
-    sendDirectMessage({
-      receiverId: dmTargetUser.id,
-      receiverNickname: dmTargetUser.nickname,
-      message,
-    })
   }
 
   function handleOpenCreateRoomModal() {
@@ -406,7 +263,7 @@ function LobbyPage() {
           isOpen={isUserListOpen}
           currentUserId={session.userId}
           unreadDirectMessageCountByUserId={unreadDirectMessageCountByUserId}
-          onOpenDirectMessage={handleOpenDirectMessage}
+          onOpenDirectMessage={openDirectMessage}
           onToggle={() => setIsUserListOpen((prev) => !prev)}
         />
       </main>
@@ -446,8 +303,8 @@ function LobbyPage() {
           user={dmTargetUser}
           currentUserId={session.userId}
           messages={directMessagesByUserId[dmTargetUser.id] ?? []}
-          onClose={handleCloseDirectMessage}
-          onSendMessage={handleSendDirectMessage}
+          onClose={closeDirectMessage}
+          onSendMessage={sendDirectMessage}
         />
       ) : null}
     </div>
